+++ agents/template/src/main/scala/org/broadinstitute/transporter/transfer/TransferRunner.scala
@@ -1,6 +1,6 @@
 package org.broadinstitute.transporter.transfer
 
-import org.broadinstitute.transporter.kafka.TransferState
+import org.broadinstitute.transporter.kafka.TransferStep
 
 /**
   * Component capable of actually running data transfers.
@@ -8,18 +8,18 @@ import org.broadinstitute.transporter.kafka.TransferState
   * Agents are expected to "fill in" an instance of this interface
   * to handle specific storage source / destination pairs.
   */
-trait TransferRunner[In, Progress, Out] {
+trait TransferRunner[I, P, O] {
 
   /**
     * Initialize the transfer described by the given request, and
     * emit enough information to push the first chunk of data.
     */
-  def initialize(request: In): TransferState[In, Progress, Out]
+  def initialize(request: I): TransferStep[I, P, O]
 
   /**
     * Push the next chunk of data into an initialized transfer, either
     * completing the transfer or emitting enough information to push the
     * following chunk.
     */
-  def step(progress: Progress): Either[Progress, Out]
+  def step(progress: P): Either[P, O]
 }
